Implemented DiskConfig

+++ tests/SakuraInternet/Saclient/ServerTest.php
@@ -12,6 +12,9 @@ class ServerTest extends \PHPUnit_Framework_TestCase
 {
 	use \SakuraInternet\Saclient\Tests\Common;
 	
+	const TESTS_DUPLICATE_DISK = false;
+	const TESTS_CONFIG_DISK = false;
+	
 	public function testAuthorize()
 	{
 		return $this->authorize();
@@ -53,6 +56,9 @@ class ServerTest extends \PHPUnit_Framework_TestCase
 		$tag = "saclient-test";
 		$cpu = 1;
 		$mem = 2;
+		$hostName = "saclient-test";
+		$sshPublicKey = "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQC3sSg8Vfxrs3eFTx3G//wMRlgqmFGxh5Ia8DZSSf2YrkZGqKbL1t2AsiUtIMwxGiEVVBc0K89lORzra7qoHQj5v5Xlcdqodgcs9nwuSeS38XWO6tXNF4a8LvKnfGS55+uzmBmVUwAztr3TIJR5TTWxZXpcxSsSEHx7nIcr31zcvosjgdxqvSokAsIgJyPQyxCxsPK8SFIsUV+aATqBCWNyp+R1jECPkd74ipEBoccnA0pYZnRhIsKNWR9phBRXIVd5jx/gK5jHqouhFWvCucUs0gwilEGwpng3b/YxrinNskpfOpMhOD9zjNU58OCoMS8MA17yqoZv59l3u16CrnrD saclient-test@local";
+		$sshPrivateKeyFile = dirname(dirname(dirname(__DIR__))) . "/test-sshkey.txt";
 		
 		// search archives
 		echo "searching archives...\n";
@@ -113,9 +119,19 @@ class ServerTest extends \PHPUnit_Framework_TestCase
 		echo "connecting the disk to the server...\n";
 		$disk->connectTo($server);
 		
+		if (self::TESTS_CONFIG_DISK) {
+			// config the disk
+			$diskconf = $disk->createConfig();
+			$diskconf->hostName = "saclient-test";
+			$diskconf->password = uniqid();
+			$diskconf->sshKey = $sshPublicKey;
+			$diskconf->write();
+		}
+		
 		// boot
 		echo "booting the server...\n";
 		$server->boot();
+		sleep(3);
 		
 		// boot conflict
 		echo "checking the server power conflicts...\n";
@@ -128,8 +144,24 @@ class ServerTest extends \PHPUnit_Framework_TestCase
 		}
 		if (!$ok) $this->fail('サーバ起動中の起動試行時は HttpConflictException がスローされなければなりません');
 		
+		if (self::TESTS_CONFIG_DISK) {
+			// ssh
+			$ipAddress = $server->ifaces[0]->ipAddress;
+			$this->assertNotEmpty($ipAddress);
+			$cmd = "ssh -oStrictHostKeyChecking=no -oUserKnownHostsFile=/dev/null -i$sshPrivateKeyFile root@$ipAddress hostname 2>/dev/null";
+			$sshSuccess = false;
+			echo "trying to SSH to the server...\n";
+			for ($i=0; $i<10; $i++) {
+				sleep(5);
+				if ($hostName != trim(`$cmd`)) continue;
+				$sshSuccess = true;
+				break;
+			}
+			$this->assertTrue($sshSuccess, '作成したサーバへ正常にSSHできません');
+		}
+		
 		// stop
-		sleep(3);
+		sleep(1);
 		echo "stopping the server...\n";
 		if (!$server->stop()->sleepUntilDown()) $this->fail('サーバが正常に停止しません');
 		
@@ -141,28 +173,31 @@ class ServerTest extends \PHPUnit_Framework_TestCase
 		echo "deleting the server...\n";
 		$server->destroy();
 		
-		// duplicate the disk
-		echo "duplicating the disk (expanding to 40GiB)...\n";
-		$disk2 = $api->disk->create();
-		$disk2->name = $name . "-copy";
-		$disk2->description = $description;
-		$disk2->tags = [$tag];
-		$disk2->source = $disk;
-		$disk2->sizeGib = 40;
-		$disk2->save();
-		
-		// wait disk duplication
-		echo "waiting disk duplication...\n";
-		if (!$disk2->sleepWhileCopying()) $this->fail("ディスクの複製がタイムアウトしました");
-		$disk2->source = null;
-		$disk2->reload();
-		$this->assertInstanceOf("SakuraInternet\\Saclient\\Cloud\\Resource\\Disk", $disk2->source);
-		$this->assertEquals($disk->id, $disk2->source->id);
-		$this->assertEquals(40, $disk2->sizeGib);
+		$disk2 = null;
+		if (self::TESTS_DUPLICATE_DISK) {
+			// duplicate the disk
+			echo "duplicating the disk (expanding to 40GiB)...\n";
+			$disk2 = $api->disk->create();
+			$disk2->name = $name . "-copy";
+			$disk2->description = $description;
+			$disk2->tags = [$tag];
+			$disk2->source = $disk;
+			$disk2->sizeGib = 40;
+			$disk2->save();
+			
+			// wait disk duplication
+			echo "waiting disk duplication...\n";
+			if (!$disk2->sleepWhileCopying()) $this->fail("ディスクの複製がタイムアウトしました");
+			$disk2->source = null;
+			$disk2->reload();
+			$this->assertInstanceOf("SakuraInternet\\Saclient\\Cloud\\Resource\\Disk", $disk2->source);
+			$this->assertEquals($disk->id, $disk2->source->id);
+			$this->assertEquals(40, $disk2->sizeGib);
+		}
 		
 		// delete the disks
 		echo "deleting the disks...\n";
-		$disk2->destroy();
+		if ($disk2) $disk2->destroy();
 		$disk->destroy();
 		
 	}
